Workaround for CIFS mounts of shares without UNIX extensions
Servers not providing UNIX extensions, or clients which wish not to
use them, face a problem with posixovl. posixovl uses lock upgrading
(converting from read locks to write locks) when updating HCBs. It is
not possible with the lack of upgrading to do locking in a race-free,
non-deadlocking way. The only feasible workaround is to do without
any locking, and that mandates single threadedness, which this patch
provides.