gprof debugging: chdir back to the original directory before exiting
The gmon.out file is written to whatever is the current working
directory, which may not be the actual directory the developer was
compiling and running ./mount.posixovl from. By changing back to the
original working directory before program exit (before gmon.out is
written), the gmon file will exist where it would be assumed.